% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charclass.R
\name{charclass}
\alias{charclass}
\title{Find Best Character Class.}
\usage{
charclass(map, enemy_class, team = "all")
}
\arguments{
\item{map}{The map name, typically beginning with "ctf_", "cp_", "pl_", "arena_", "plr_", or "koth_"}

\item{enemy_class}{One of scout, sniper, soldier, demoman, medic, heavy, pyro, spy, or engineer}

\item{team}{(Optional) BLU (offence) or RED (defence)}
}
\value{
A bar chart displaying the number of kills each class made for the given map and enemy class.
}
\description{
This function helps players choose the best character class to KILL a given enemy.
}
\details{
This function helps players find the best class for killing a given enemy class.
This is a function of map, enemy’s class, and optionally team.
For asymmetric maps (pl_ and some cp_), BLU represents the attacking team and RED the defending team.
}
\examples{
charclass("koth_viaduct", enemy_class = "sniper")
charclass("pl_goldrush", enemy_class = "spy", team = "RED")
}
