% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{elevation}
\alias{elevation}
\title{Find Weapon Elevation Statistics.}
\usage{
elevation(map, weapon, enemy_class = NULL)
}
\arguments{
\item{map}{The map name, typically beginning with "ctf_", "cp_", "pl_", "arena_", "plr_", or "koth_"}

\item{weapon}{An integer referring to the weapon from the TF2 item index at https://wiki.alliedmods.net/Team_Fortress_2_Item_Definition_Indexes}

\item{enemy_class}{(Optional) One of scout, sniper, soldier, demoman, medic, heavy, pyro, spy, or engineer}
}
\value{
A kernel density estimate of probability of kills by elevation.
}
\description{
This function shows how elevation affects kills.
}
\details{
This function shows the distribution of elevation differences between killers and victims, by the map, killers' weapon, and optionally the enemy class.
If no enemy class is specified, then the function will use kills against all enemy classes.
Elevation is defined such that it is positive if the killer is on higher ground than the victim.
Function may take several seconds to run depending on map.
}
\examples{
elevation("pl_goldrush", 13) ## scattergun (short range)
elevation("pl_goldrush", 14) ## sniper rifle (long range)
elevation("pl_goldrush", 14, "sniper") ## against enemy snipers
elevation("ctf_doublecross", 15150) ## returns message if no kills for the parameters entered
}
