% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmap.R
\name{findmap}
\alias{findmap}
\title{Find TF2 Map.}
\usage{
findmap(pattern, minkills = 1)
}
\arguments{
\item{pattern}{A regex pattern to be used to search for maps}

\item{minkills}{The threshold for minimum number of kills a map must have to be included}
}
\value{
A dataframe of matching map names and their associated kill counts
}
\description{
This function searches for a map using a regex pattern.
}
\details{
This function searches for map by a regex pattern, and returns matching maps with the number of kills recorded for each map.
Maps with fewer kills than the threshold set are automatically filtered out. The threshold is 1 by default.
}
\examples{
findmap("plr_")
findmap("ctf_", 900)
findmap("event$")
findmap("(oo).*\\\\1")
}
