% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weprange.R
\name{weprange}
\alias{weprange}
\title{Find Weapon Range Statistics.}
\usage{
weprange(map, weapon, enemy_class = NULL, distance = "3D")
}
\arguments{
\item{map}{The map name, typically beginning with "ctf_", "cp_", "pl_", "arena_", "plr_", or "koth_"}

\item{weapon}{An integer referring to the weapon from the TF2 item index at https://wiki.alliedmods.net/Team_Fortress_2_Item_Definition_Indexes}

\item{enemy_class}{(Optional) One of scout, sniper, soldier, demoman, medic, heavy, pyro, spy, or engineer}

\item{distance}{Either "2D" or "3D" (default)}
}
\value{
A kernel density estimate of probability of kills by range.
}
\description{
This function shows how weapon range affects kills.
}
\details{
This function shows the distribution of kill ranges, by the map, killers' weapon, and optionally the enemy class.
If no enemy class is specified, then the function will use kills against all enemy classes.
Range can be set to either 2D (only x-y) or 3D distance.
Function may take several seconds to run depending on the map.
}
\examples{
weprange("pl_goldrush", 21) ## flamethrower (short range)
weprange("pl_goldrush", 56) ## huntsman (long range)
weprange("pl_goldrush", 56, enemy_class="sniper", distance = "2D")
weprange("ctf_doublecross", 15150) ## message if no kills for the parameters entered
}
